+++ src/template/Roadmaps/heroRoadmap/index.tsx
@@ -1,10 +1,44 @@
 import { HeroTitleRoadmaps, StrongTitle } from "./styled";
 import { Grid } from "../../../components/Grid";
 import { useNavigate } from "react-router-dom";
-import { buttonsRoadmaps } from "../GridButton/data";
+import { useEffect, useState } from "react";
+import { ListingRoadmapsApi } from "../../../services/roadmap/listing/api";
+
+
+interface Roadmap {
+   id: number | null;
+   category: string;
+ 
+ }
 
 export function HeroRoadmaps() {
   const navigate = useNavigate();
+
+  const [roadmaps, setRoadmaps] = useState<Roadmap[]>([]);
+
+   async function fetchRoadmap() {
+       try {
+         const data = await ListingRoadmapsApi();
+         setRoadmaps(data);
+       } catch (error) {
+         console.error("Error fetching roadmaps:", error);
+       }
+     }
+     
+useEffect(() => {
+  fetchRoadmap()
+})
+
+  const handleNavigateBackend = () => {
+    roadmaps.map((roadmap) => {
+    if(roadmap.category.toLocaleLowerCase() === "back-end" || roadmap.category.toLocaleLowerCase() === "backend") {
+      navigate("/roadmaps/back-end");
+    }else{
+      navigate("/roadmaps")
+    }
+    return roadmap.category;
+  })
+}
 
   return (
     <>
@@ -15,9 +49,13 @@ export function HeroRoadmaps() {
       <Grid
         columns={3}
         gap={"32px"}
-        navigate={(index) => navigate(buttonsRoadmaps[index].path)}
+        navigate={handleNavigateBackend}
+        
       >
-        {buttonsRoadmaps.map((button) => button.name)}
+        
+        {roadmaps.map((roadmap) =>roadmap.category)}
+        
+      
       </Grid>
     </>
   );
